+++ ChatApp/ChatModule.DBShema/Models/User.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.ComponentModel.DataAnnotations;
 
 namespace ChatModule.DBShema.Models
@@ -12,7 +13,9 @@ namespace ChatModule.DBShema.Models
         public string last_name { get; set; }
         public string avatarlink { get; set; }
         public DateTime? last_logout { get; set; }
+        [DefaultValue("offline")]
+        public string status { get; set; }
         //For Entitys Framework's "Lazy Loading" ->  Lazy Loading means that the contents of these properties will be automatically loaded from the database when you try to access them
-        public virtual ICollection<Category2User> Category2User { get; set; }
+        public virtual ICollection<User2SupportGroup> Category2User { get; set; }
     }
 }
